Added test methods for login and register
Tests when a user logs in with credentials that are registered and unregistered 

Also tests that a user is registered correctly with the register view. Tests to make sure duplicate users cannot be added, but the code for this is broken. Something to do with the IntegrityError exception, I'll try and fix